Enum bzip2::CompressionLevel

pub enum CompressionLevel {
    Fastest,
    Smallest,
    Default,
}

When compressing data, the compression level can be specified by a value in this enum.

Variants

Fastest

Optimize for the best speed of encoding.

Smallest

Optimize for the size of data being encoded.

Default

Choose the default compression, a balance between speed and size.