Struct bzip2::Decompress

pub struct Decompress { /* fields omitted */ }

Representation of an in-memory decompression stream.

An instance of Decompress can be used to inflate a straem of bz2-encoded data.

Methods

impl Decompress

fn new(small: bool) -> Decompress

Creates a new stream prepared for decompression.

If small is true, then the library will use an alternative decompression algorithm which uses less memory but at the cost of decompressing more slowly (roughly speaking, half the speed, but the maximum memory requirement drops to around 2300k). See

fn decompress(&mut self,
input: &[u8],
output: &mut [u8])
-> Result<Status, Error>

Decompress a block of input into a block of output.

fn decompress_vec(&mut self,
input: &[u8],
output: &mut Vec<u8>)
-> Result<Status, Error>

Decompress a block of input into an output vector.

This function will not grow output, but it will fill the space after its current length up to its capacity. The length of the vector will be adjusted appropriately.

fn total_in(&self) -> u64

Total number of bytes processed as input

fn total_out(&self) -> u64

Total number of bytes processed as output