Struct bzip2::write::BzDecoder

pub struct BzDecoder<W: Write> { /* fields omitted */ }

A compression stream which will have compressed data written to it and will write uncompressed data to an output stream.

Methods

impl<W: Write> BzDecoder<W>

Important traits for BzDecoder<W>
impl<W: Read + Write> Read for BzDecoder<W>impl<W: Write> Write for BzDecoder<W>

pub fn new(obj: W) -> BzDecoder<W>

Create a new decoding stream which will decompress all data written to it into obj.

pub fn get_ref(&self) -> &W

Acquires a reference to the underlying writer.

pub fn get_mut(&mut self) -> &mut W

Acquires a mutable reference to the underlying writer.

Note that mutating the output/input state of the stream may corrupt this object, so care must be taken when using this method.

pub fn try_finish(&mut self) -> Result<()>

Attempt to finish this output stream, writing out final chunks of data.

Note that this function can only be used once data has finished being written to the output stream. After this function is called then further calls to write may result in a panic.

Panics

Attempts to write data to this stream may result in a panic after this function is called.

pub fn finish(&mut self) -> Result<W>

Unwrap the underlying writer, finishing the compression stream.

Note that this function may not be suitable to call in a situation where the underlying stream is an asynchronous I/O stream. To finish a stream the try_finish (or shutdown) method should be used instead. To re-acquire ownership of a stream it is safe to call this method after try_finish or shutdown has returned Ok.

pub fn total_out(&self) -> u64

Returns the number of bytes produced by the decompressor

Note that, due to buffering, this only bears any relation to total_in() after a call to flush(). At that point, total_in() / total_out() is the compression ratio.

pub fn total_in(&self) -> u64

Returns the number of bytes consumed by the decompressor (e.g. the number of bytes written to this stream.)