Struct bzip2::write::BzEncoder

pub struct BzEncoder<W: Write> {
    // some fields omitted
}

A compression stream which will have uncompressed data written to it and will write compressed data to an output stream.

Methods

impl<W: Write> BzEncoder<W>

fn new(obj: W, level: Compression) -> BzEncoder<W>

Create a new compression stream which will compress at the given level to write compress output to the give output stream.

fn finish(self) -> Result<W>

Consumes this encoder, flushing the output stream.

This will flush the underlying data stream and then return the contained writer if the flush succeeded.

fn total_out(&self) -> u64

Returns the number of bytes produced by the compressor

Note that, due to buffering, this only bears any relation to total_in() after a call to flush(). At that point, total_out() / total_in() is the compression ratio.

fn total_in(&self) -> u64

Returns the number of bytes consumed by the compressor (e.g. the number of bytes written to this stream.)