Struct bzip2::writer::BzDecompressor

pub struct BzDecompressor<W: Write> {
    // some fields omitted
}

A compression stream which will have compressed data written to it and will write uncompressed data to an output stream.

Methods

impl<W: Write> BzDecompressor<W>

fn new(w: W) -> BzDecompressor<W>

Create a new compression stream which will compress at the given level to write compress output to the give output stream.

fn into_inner(self) -> Result<W, (BzDecompressor<W>, Error)>

Unwrap the underlying writer, finishing the compression stream.